tests: Do not overwrite test log files
zonefs-tests-nullblk.sh execute 3 series of tests using nullblk devices
with a different number of conventional zones. Since for each case the
device name is the same, the run log file is always overwritten and only
the log file for the last run remains.

Avoid this by adding the number of conventional zones to the log file
and allowing to specify a log file with the -g option in the
zonefs-tests.sh script.

Signed-off-by: Masato Suzuki <masato.suzuki@wdc.com>
Signed-off-by: damien Le Moal <damien.lemoal@wdc.com>