tests: Fix file size checks
In test 0086 to 0089, test the file size after writing against the file
zone capacity instead of the zone size. To do this, introduce the
variables seq_file_0_zone_start_sector and seq_file_0_max_size to
indicate the start sector and capacity of the zone for the file seq/0
used in these tests.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>