mkzonefs: Always reset all zones
When mkzonefs formats a file system on a SMR drive, write pointers on
the drive can be inconsistent with those tracked by SCSI driver because
of pass-through commands issued to the drive by other applications.

To ensure that all write pointers are consistent at file system
creation time, always reset all zones regardless of their conditions.
Modify zonefs_reset_zones() to call reset zone ioctl for all zones.
This all zones reset is fast but fails if the target device is
a device-mapper device with conventional zones. In such a case, fallback
to slow path to call zonefs_reset_zone() for each zone. Also modify
zonefs_reset_zone() to reset the zone even when it is empty and make
that function static.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>