tests: Fix stale data access
Reseting a zone using blkzone reset does not invalidate the zone data
present in the page cache. This can cause false positive from libblkid
probe and test failures. When the super block is in a sequential zone
and cleared with a zone reset, drop inode pages from the page cache to
avoid this problem.

Suggested-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>