dev: refine reset all zones handling
There is no point in trying a reset all zones operation for devices
known to not support this operation. Introduce the helper function
zonefs_dev_has_reset_all() to detect support for reset all operation.
For now, only exclude DM devices which are know to not support reset
all, regardless of the target used.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>