tests: 0514: Fix max open handling
A device may have no limit on the number of open zones. In such case, do
not test for open failures with explicit_open mount option, but only and
only if the device also has no active zone limit. For a device that haz
max open equal to 0 but a limit on the number of active zones, that
limit should be used as the maximum number of files that can be open.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>