tests: test_lib: fix test failures with symbolic links to block devices
When zonefs-tests.sh takes a symbolic link as the test target zoned
device, test cases 512, 513, 514 and 515 fail. Cause of the failures is
that device name for sysfs path was extracted from the symbolic link
name instead of the resolved real name of the device. The failures are
observed with device-mapper symbolic links in /dev/mapper.

To avoid the failures, resolve the symbolic links before getting sysfs
path of the test target devices. Add a new helper function devname() to
implement it.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@opensource.wdc.com>