tests: 0215: avoid write beyond conventional file size
The test case 0215 writes data to the aggregated conventional zone file
with 16MB block size. The count of write blocks is calculated by
dividing the file size by the block size. When the file size is not a
multiple of 16MB, the block count is incremented so that the file is
written up to its maximum size. However, this cause write beyond file
size limit and "File too large" error is reported. To avoid the error,
do not increment the write block count.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>