tests: Allow tests to run with built-in zonefs
If the test kernel has zonefs built-in (not in a module), running
zonefs_tests.sh results in all tests failed. This happens because the
test suite loads and unloads the zonefs kernel module before/after every
individual test. This is an absolutely right thing to do for catching
module reference leaks, but if zonefs is built-in, the rmmod that is
executed after every test fails with "rmmod: ERROR: Module zonefs is
builtin." error and this makes the test fail.

To fix, check if zonefs is built as a kernel module and bypass all
modprobe/rmmod invocations during the test run if there is no module
called zonefs. If this is the case, verify that the kernel supports
zonefs file system at all by looking at /proc/filesystems and don't
run the tests if zonefs is not supported.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>