tests: Fix setting device I/O scheduler
Since Linux kernel version v6.5, the sysfs attribute "queue/scheduler"
is no longer available for bio based device-mapper targets. When
zonefs-tests.sh runs for such device, it fails to set up mq-deadline
as the device scheduler, which is required for zoned block devices.

To ensure that mq-deadline scheduler is applied during zonefs-tests.sh
run for bio based device-mapper targets, set mq-deadline scheduler for
the destination devices of the device-mapper device. If the sysfs
attribute "queue/scheduler" is not available, check if the test device
is a zoned device-mapper (linear, flakey or crypt). If so, get the
destination devices and set the I/O scheduler for these devices.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>