tests: Fix the exit status of zonefs-tests.sh
It is expected that tests/zonefs-tests.sh exits with a non-zero status
code when any of the test cases fail. However, it always returns zero
status after the commit 18cb880fbec8 ("tests: log zonefs-tests.sh
output"). The commit introduced curly brackets to pass outputs to a tee
command. It made bash variable changes in the brackets invalid for the
following scripts. Hence, the variable "failed" was not updated
correctly and the status code was always zero.

Fix this by checking the file that the tee command outputs and update
the "failed" variable with the correct value.

Fixes: 18cb880fbec8 ("tests: log zonefs-tests.sh output")
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>