tests: 0002: Fix missing library source command and mkfs.f2fs failures
The test case does not source scripts/test_lib then exit_failed() bash
function defined in scripts/test_lib is not working. The test case was
not able to report test failure correctly, and the following issues were
hidden. Source the script library to not hide failures.

On top of that, this test case fails at the mkfs.f2fs -m command when

 1) the test target device has a filesystem,
 2) the test target device does not have conventional zones, or
 3) the test target device size is larger than 16TB (f2fs design limit).

To avoid the failures, add the -f option to mkfs.f2fs, check
conventional zones existence and specify filesystem size up to 16TB.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>