tests: Increase scsi_debug device size
The test script zonefs-tests-scsi-debug.sh prepares 1GB scsi_debug
device for testing. However, the size is not enough mkfs.f2fs command
and causes failure of the test case 0002. To avoid the failure, increase
the size to 2GB. This size is same as null_blk device size that
zonefs-tests-nullblk.sh prepares.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>