tests: 0304, 0305: Fix error test
Tests 0304 and 0305 succeed as dd and fio (respectively) fail to write
the sequential file as expected. However, the reason for the failure is
that oflag=direct is missing for test 0304 and fio --size option is
missing for test 0305.

Fix 0304 test case by adding the missing oflag=direct option. Also check
that the file is still writable after checking for the initial write
error and file size.

Fix 0305 by replacing the call to fio with zio, and also check
that the file is still writable after checking for the initial write
error and file size.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>