tests: fix the 5th drive configuration in zonefs-tests-nullblk.sh
Commit eb45c3fb3ece ("tests: improve zonefs-tests-nullblk.sh") added the
5th drive configuration to set up the null_blk device with zone capacity
smaller than zone size. The total null_blk device size of the
configuration was succeeded from the 4th drive configuration, which is
not multiple of the zone size. Then the last zone is smaller than the
other zones.

On the other hand, after the kernel commit b164316808ec ("null_blk: Do
not allow runt zone with zone capacity smaller then zone size"), it is
no longer allowed to create null_blk such as "A zoned device with a
smaller last zone together with a zone capacity smaller than the zone
size". Then, zonefs-tests-nullblk.sh fails to set up the 5th drive
configuration.

To avoid the failure, make the total device size of the 5th drive
configuration a multiple of the zone size.

Fixes: eb45c3fb3ece ("tests: improve zonefs-tests-nullblk.sh")
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <dlemoal@kernel.org>